# Speedup of canPassThrough() method